Replace `Vec` and `String` with `FixedArray` and `FixedString` for all models (#2656)
This shrinks type sizes by a lot; however, it makes the user experience slightly
different:

- `FixedString` must be converted to String with `.into()` or `.into_string()`
  before it can be pushed to, but dereferences to `&str` as is.
- `FixedArray` must be converted to `Vec` with `.into()` or `.into_vec()`
  before it can be pushed to, but dereferences to `&[T]` as is.

The crate of these types is currently a Git dependency, but this is fine for
the `next` branch. It needs some basic testing, which Serenity is perfect for,
before a release will be made to crates.io.

+++ src/framework/standard/parse/mod.rs
@@ -55,7 +55,7 @@ fn permissions_in(
     }
 
     if let Some(channel) = guild.and_then(|guild| guild.channels.get(&channel_id).cloned()) {
-        let mut data = Vec::with_capacity(member.roles.len());
+        let mut data = Vec::with_capacity(member.roles.len() as usize);
 
         for overwrite in &channel.permission_overwrites {
             if let PermissionOverwriteType::Role(role) = overwrite.kind {
